requirements.txt
pydantic
fastapi[standard]
pytest
pandas
openai
psycopg
python-dotenv
timescale-vector
instructor
anthropic
uvicorn